Add error handling for Question Table
Cleaned up file tree and placed depricated code the Questions module
into `Quesitions (prev)` in their original root folders.
