+++ client/src/components/Questions (prev)/Write/QuestionExample.tsx
