+++ client/src/components/Questions (prev)/Write/QuestionFooter.tsx
