+++ client/src/components/Questions (prev)/Write/SectionHeader.tsx
