+++ client/src/components/Questions/Read/QuestionConstraints.tsx
@@ -1,26 +1,20 @@
-import { ArrowRightIcon } from '@heroicons/react/24/outline';
-import { QuestionConstraint as QuestionConstraintType } from '../../../types';
-import SectionHeader from './SectionHeader';
+import { useSelector } from 'react-redux';
+import { selectConstraints } from '../../../features/questions/creatorSlice';
 
-interface QuestionConstraintsProps {
-  constraints?: QuestionConstraintType[];
-}
+const QuestionConstraints = () => {
+  const constraints = useSelector(selectConstraints);
 
-const QuestionConstraints: React.FC<QuestionConstraintsProps> = ({
-  constraints
-}) => {
   return (
-    <div className="flex flex-col w-full gap-2">
-      <SectionHeader title="Constraints" />
-
-      <div className="flex flex-col gap-1 text-gray-100 text-sm">
-        {constraints?.map((constraint, index) => (
-          <div key={index} className="flex gap-4 items-center">
-            <ArrowRightIcon className="w-4 h-4" />
-            {constraint}
-          </div>
+    <div
+      className="flex flex-col
+        gap-2"
+    >
+      <h3 className="font-semibold">Constraints</h3>
+      <ul className="list-disc list-inside text-gray-700">
+        {constraints.map((constraint, index) => (
+          <li key={index}>{constraint}</li>
         ))}
-      </div>
+      </ul>
     </div>
   );
 };
