+++ client/src/components/Questions/Write/QuestionConstraints.tsx
@@ -1,6 +1,3 @@
-import { ArrowRightIcon } from '@heroicons/react/24/outline';
-import { MinusIcon, PlusIcon } from '@heroicons/react/24/solid';
-import { useCallback } from 'react';
 import { useSelector } from 'react-redux';
 import {
   addConstraint,
@@ -8,56 +5,69 @@ import {
   selectConstraints,
   updateConstraint
 } from '../../../features/questions/creatorSlice';
+import { PlusIcon, XCircleIcon } from '@heroicons/react/24/solid';
 import { store } from '../../../store';
-import { QuestionConstraint } from '../../../types';
-import SectionHeader from './SectionHeader';
 
 const QuestionConstraints = () => {
   const constraints = useSelector(selectConstraints);
 
-  const onAdd = useCallback(() => {
+  const handleDelete = (
+    e: React.MouseEvent<SVGSVGElement, MouseEvent>,
+    index: number
+  ) => {
+    e.preventDefault();
+    store.dispatch(
+      deleteConstraint({ constraint: constraints[index], index: index })
+    );
+  };
+
+  const handleChange = (
+    e: React.ChangeEvent<HTMLInputElement>,
+    index: number
+  ) => {
+    e.preventDefault();
+    store.dispatch(
+      updateConstraint({ constraint: e.target.value, index: index })
+    );
+  };
+
+  const handleAddConstraint = (
+    e: React.MouseEvent<HTMLButtonElement, MouseEvent>
+  ) => {
+    e.preventDefault();
     store.dispatch(addConstraint(''));
-  }, [store, constraints]);
-
-  const onRemove = useCallback(
-    (constraint: QuestionConstraint) => {
-      store.dispatch(deleteConstraint(constraint));
-    },
-    [store, constraints]
-  );
-
-  const handleChange = useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>, index: number) => {
-      store.dispatch(updateConstraint({ constraint: e.target.value, index }));
-    },
-    [store]
-  );
+  };
 
   return (
-    <div className="flex flex-col bg-gray-800 p-8 rounded-2xl shadow-lg gap-4">
-      <SectionHeader title="Constraints" />
-
+    <div
+      className="flex flex-col
+        gap-2"
+    >
+      <h3 className="font-semibold">Constraints</h3>
       {constraints.map((constraint, index) => (
-        <div key={index} className="flex gap-4 items-center px-4 text-gray-100">
-          <ArrowRightIcon className="w-4 h-4" />
+        <div
+          key={index}
+          className="relative border border-[0.5] bg-white p-4 rounded-xl"
+        >
+          <XCircleIcon
+            className="absolute top-2 right-2 w-5 h-5 text-red-500 hover:cursor-pointer"
+            onClick={(e) => handleDelete(e, index)}
+          />
           <input
+            key={index}
             type="text"
+            className="w-full border border-gray-300 rounded-md p-2"
+            placeholder="Enter constraint"
             value={constraint}
             onChange={(e) => handleChange(e, index)}
-            className="rounded-md outline-none px-2 border shadow-inner bg-gray-800 w-full"
           />
-          <button onClick={() => onRemove(constraint)} className="pl-3 flex">
-            <div className="p-1 bg-gray-700 rounded-full shadow-lg hover:scale-105">
-              <MinusIcon className="w-4 h-4 text-rose-500" />
-            </div>
-          </button>
         </div>
       ))}
-
-      <button onClick={onAdd} className="pl-3 flex">
-        <div className="p-1 bg-gray-700 rounded-full shadow-lg hover:scale-105">
-          <PlusIcon className="w-4 h-4 text-gray-100" />
-        </div>
+      <button
+        onClick={(e) => handleAddConstraint(e)}
+        className="bg-blue-500 text-white px-4 py-2 rounded-xl"
+      >
+        <PlusIcon className="h-5 w-5 inline-block" />
       </button>
     </div>
   );
