+++ client/src/features/questions/creatorSlice.ts
@@ -40,7 +40,6 @@ axios.defaults.withCredentials = true;
 export const createQuestion = createAsyncThunk(
   '/creatorSlice/createQuestion',
   async (question: Question) => {
-    console.log(question);
     const response = await axios.post('/question-api/questions', question);
     return response.data;
   }
