+++ client/src/modules/Play/Play.tsx
@@ -1,10 +1,10 @@
 import { useSelector } from 'react-redux';
 import Editor from '../../components/Editor/Editor';
 import PlayBox from '../../components/Playbox/PlayBox';
-import Question from '../../components/Questions/Question';
+import Question from '../../components/Questions (prev)/Question';
 import { selectIsActive } from '../../features/play/playSlice';
-import { ToastContainer } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
+import { ToastContainer } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 
 const Play = () => {
   const isActive = useSelector(selectIsActive);
